+++ ui/src/main/java/ru/tinkoff/acquiring/sdk/TinkoffAcquiring.kt
@@ -22,14 +22,21 @@ import ru.tinkoff.acquiring.sdk.models.AsdkState
 import ru.tinkoff.acquiring.sdk.models.CollectDataState
 import ru.tinkoff.acquiring.sdk.models.DefaultState
 import ru.tinkoff.acquiring.sdk.models.PaymentSource
+import ru.tinkoff.acquiring.sdk.models.options.FeaturesOptions
 import ru.tinkoff.acquiring.sdk.models.options.screen.AttachCardOptions
 import ru.tinkoff.acquiring.sdk.models.options.screen.BaseAcquiringOptions
 import ru.tinkoff.acquiring.sdk.models.options.screen.PaymentOptions
+import ru.tinkoff.acquiring.sdk.models.options.screen.SavedCardsOptions
 import ru.tinkoff.acquiring.sdk.models.paysources.AttachedCard
 import ru.tinkoff.acquiring.sdk.models.paysources.CardData
 import ru.tinkoff.acquiring.sdk.models.paysources.GooglePay
 import ru.tinkoff.acquiring.sdk.payment.PaymentProcess
-import ru.tinkoff.acquiring.sdk.ui.activities.*
+import ru.tinkoff.acquiring.sdk.ui.activities.AttachCardActivity
+import ru.tinkoff.acquiring.sdk.ui.activities.BaseAcquiringActivity
+import ru.tinkoff.acquiring.sdk.ui.activities.PaymentActivity
+import ru.tinkoff.acquiring.sdk.ui.activities.SavedCardsActivity
+import ru.tinkoff.acquiring.sdk.ui.activities.StaticQrActivity
+import ru.tinkoff.acquiring.sdk.ui.activities.ThreeDsActivity
 
 /**
  * Точка входа для взаимодействия с Acquiring SDK
@@ -136,18 +143,46 @@ class TinkoffAcquiring(
     }
 
     /**
-     * Запуск экрана с отображением QR кода для оплаты покупателем
+     * Запуск экрана Acquiring SDK для просмотра сохраненных карт
+     *
+     * @param activity          контекст для запуска экрана
+     * @param savedCardsOptions настройки экрана сохраненных карт
+     * @param requestCode       код для получения результата, по завершению работы экрана Acquiring SDK
      */
-    fun openStaticQrScreen(activity: FragmentActivity, localization: LocalizationSource, requestCode: Int) {
+    fun openSavedCardsScreen(activity: FragmentActivity, savedCardsOptions: SavedCardsOptions, requestCode: Int) {
+        savedCardsOptions.setTerminalParams(terminalKey, password, publicKey)
+        val intent = BaseAcquiringActivity.createIntent(activity, savedCardsOptions, SavedCardsActivity::class.java)
+        activity.startActivityForResult(intent, requestCode)
+    }
+
+    /**
+     * Запуск экрана с отображением QR кода для оплаты покупателем
+     *
+     * @param activity        контекст для запуска экрана
+     * @param featuresOptions конфигурация визуального отображения экрана
+     * @param requestCode     код для получения результата, по завершению работы экрана Acquiring SDK
+     */
+    fun openStaticQrScreen(activity: FragmentActivity, featuresOptions: FeaturesOptions, requestCode: Int) {
         val options = BaseAcquiringOptions().apply {
             setTerminalParams(
                     this@TinkoffAcquiring.terminalKey,
                     this@TinkoffAcquiring.password,
                     this@TinkoffAcquiring.publicKey)
-            features.localizationSource = localization
+            features = featuresOptions
         }
         val intent = BaseAcquiringActivity.createIntent(activity, options, StaticQrActivity::class.java)
         activity.startActivityForResult(intent, requestCode)
+    }
+
+    /**
+     * Запуск экрана с отображением QR кода для оплаты покупателем
+     *
+     * @param activity     контекст для запуска экрана
+     * @param localization локализация экрана
+     * @param requestCode  код для получения результата, по завершению работы экрана Acquiring SDK
+     */
+    fun openStaticQrScreen(activity: FragmentActivity, localization: LocalizationSource, requestCode: Int) {
+        openStaticQrScreen(activity, FeaturesOptions().apply { localizationSource = localization }, requestCode)
     }
 
     companion object {